+++ packages/zosfiles/src/api/methods/create/index.ts
@@ -10,6 +10,7 @@
 */
 
 export * from "./doc/ICreateDataSetOptions";
+export * from "./doc/ICreateDataSetLikeOptions";
 export * from "./doc/ICreateVsamOptions";
 
 export * from "./Create";